Add command "shuffle" for qtest
Implemented a "shuffle" command in qtest that shuffles the nodes in the
queue by calling the q_shuffle function from queue.c.

Note: The q_shuffle function needs to be declared in queue.h to be
properly used by qtest. As modification of header files is restricted,
the "shuffle" command has been commented out to avoid compilation
errors.

Change-Id: I3354177aa16c8b74f716ccc23c71cc30336c8c87